# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agri System API Contract
  version: 1.0.0
  description: |
    Frontend integration contract generated from current Spring Boot + MyBatis-Plus controllers/services.

    Unified response wrapper:
    R<T> = { code:number, msg:string, data:T }
    - business success: code = 200
    - business failure: code != 200 (common: 500)

    Pagination response shape:
    data = { current:number, size:number, total:number, records:T[] }

    Authentication:
    Authorization: Bearer <token>
servers:
  - url: http://localhost:8080
...
paths:
...
  /task/list:
    get:
      tags: [Task]
      summary: List tasks (paged)
      operationId: listTasks
      parameters:
        - in: query
          name: pageNum
          schema:
            type: integer
            default: 1
          example: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
          example: 10
        - in: query
          name: taskName
          schema:
            type: string
          example: Irrigate
        - in: query
          name: status
          schema:
            type: integer
            description: 0 pending_assign, 1 pending_accept, 2 accepted, 3 completed, 4 overdue, 5 rejected
          example: 0
        - in: query
          name: executorId
          schema:
            type: integer
            format: int64
          example: 3
        - in: query
          name: assigneeId
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: R<Page<AgriTask>>
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageAgriTask'
              example:
                code: 200
                msg: Operation success
                data:
                  current: 1
                  size: 10
                  total: 1
                  records:
                    - taskId: 1
                      batchId: 1
                      taskName: Spring Fertilization
                      taskType: Fertilize
                      priority: 1
                      planTime: "2026-03-05T08:00:00"
                      status: 1
                      executorId: 3
                      executorName: Zhang San
                      executorRoleKey: FARMER
                      executorRoleName: 农户
                      assigneeId: 3
                      assigneeName: Zhang San
                      assignTime: "2026-03-01T09:00:00"
                      assignBy: 1
                      assignRemark: Please handle today
                      acceptTime: null
                      acceptBy: null
                      rejectTime: null
                      rejectBy: null
                      rejectReason: null
                      createBy: 1
                      createTime: "2026-02-28T09:00:00"
                      updateBy: null
                      updateTime: null
                      version: 0
        '500':
          $ref: '#/components/responses/BizFail'
  /task:
    post:
      tags: [Task]
      summary: Create task (status forced to 0)
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgriTaskCreate'
            example:
              batchId: 1
              taskName: New Irrigation Task
              taskType: Irrigate
              priority: 2
              planTime: "2026-03-10T09:00:00"
              createBy: 1
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
    put:
      tags: [Task]
      summary: Update task basic info only
      description: Only basic fields are allowed. Status/assignee fields are not exposed by this endpoint.
      operationId: updateTaskBasic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateDTO'
            example:
              taskId: 1
              batchId: 1
              taskName: Updated Task Name
              taskType: Fertilize
              priority: 1
              planTime: "2026-03-06T08:30:00"
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
  /task/assign:
    put:
      tags: [Task]
      summary: Assign task (ADMIN/FARM_OWNER, FARMER:0->1, WORKER:0->2)
      description: |
        Closed-loop dispatch endpoint:
        - role required: ADMIN or FARM_OWNER
        - target user must be active FARMER or WORKER
        - target role mismatch returns business error: 仅支持派给 FARMER/WORKER
        - status transition:
          - assign to FARMER: pending_assign(0) -> pending_accept(1)
          - assign to WORKER: pending_assign(0) -> accepted(2)
      operationId: assignTask
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Trace-Id
          schema:
            type: string
          required: false
          example: trace-assign-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskAssignDTO'
            example:
              taskId: 1
              executorId: 3
              remark: Please handle today
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/BizFail'
  /task/accept:
    post:
      tags: [Task]
      summary: Accept task (FARMER assignee only, 1 -> 2)
      description: |
        Closed-loop dispatch endpoint:
        - role required: FARMER
        - must be assigned farmer
        - status transition: pending_accept(1) -> accepted(2)
        - WORKER calling this endpoint returns business error:
          WORKER无需接单/拒单，请直接执行任务
      operationId: acceptTask
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Trace-Id
          schema:
            type: string
          required: false
          example: trace-accept-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskAcceptDTO'
            example:
              taskId: 1
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/BizFail'
  /task/reject:
    post:
      tags: [Task]
      summary: Reject task (FARMER assignee only, 1 -> 0)
      description: |
        Closed-loop dispatch endpoint:
        - role required: FARMER
        - must be assigned farmer
        - status transition: pending_accept(1) -> pending_assign(0)
        - WORKER calling this endpoint returns business error:
          WORKER无需接单/拒单，请直接执行任务
      operationId: rejectTask
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Trace-Id
          schema:
            type: string
          required: false
          example: trace-reject-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRejectDTO'
            example:
              taskId: 1
              reason: Cannot finish today
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/BizFail'
  /task/{ids}:
    delete:
      tags: [Task]
      summary: Delete tasks by CSV ids
      operationId: deleteTask
      parameters:
        - in: path
          name: ids
          required: true
          schema:
            type: string
          example: 1,2
      responses:
        '200':
          description: R<Void>
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVoid'
              example:
                code: 200
                msg: Operation success
                data: null
        '500':
          $ref: '#/components/responses/BizFail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BizFail:
      description: Business failure (R<Void>, code != 200)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RVoid'
          example:
            code: 500
            msg: Operation failed
            data: null
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RVoid'
          example:
            code: 401
            msg: Please login first
            data: null
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RVoid'
          example:
            code: 400
            msg: taskId 和 executorId 不能为空
            data: null
  schemas:
    R:
      type: object
      properties:
        code:
          type: integer
          description: Business code, 200 success, non-200 fail
        msg:
          type: string
        data:
          nullable: true
      required: [code, msg, data]
    RVoid:
      allOf:
        - $ref: '#/components/schemas/R'
        - type: object
          properties:
            data:
              nullable: true
              example: null
    PageBase:
      type: object
      properties:
        current:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
      required: [current, size, total]
...
    PageAgriTask:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/AgriTask'
          required: [records]
...
    AgriTask:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        taskName:
          type: string
        taskType:
          type: string
        priority:
          type: integer
        planTime:
          type: string
          format: date-time
        status:
          type: integer
          description: 0 pending_assign, 1 pending_accept, 2 accepted, 3 completed, 4 overdue, 5 rejected
        executorId:
          type: integer
          format: int64
          nullable: true
        executorName:
          type: string
          nullable: true
        executorRoleKey:
          type: string
          nullable: true
          description: FARMER or WORKER
        executorRoleName:
          type: string
          nullable: true
        assigneeId:
          type: integer
          format: int64
          nullable: true
        assigneeName:
          type: string
          nullable: true
        assignTime:
          type: string
          format: date-time
          nullable: true
        assignBy:
          type: integer
          format: int64
          nullable: true
        assignRemark:
          type: string
          nullable: true
        acceptTime:
          type: string
          format: date-time
          nullable: true
        acceptBy:
          type: integer
          format: int64
          nullable: true
        rejectTime:
          type: string
          format: date-time
          nullable: true
        rejectBy:
          type: integer
          format: int64
          nullable: true
        rejectReason:
          type: string
          nullable: true
        createBy:
          type: integer
          format: int64
          nullable: true
        createTime:
          type: string
          format: date-time
        updateBy:
          type: integer
          format: int64
          nullable: true
        updateTime:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
    AgriTaskCreate:
      type: object
      properties:
        batchId:
          type: integer
          format: int64
        taskName:
          type: string
        taskType:
          type: string
        priority:
          type: integer
        planTime:
          type: string
          format: date-time
        createBy:
          type: integer
          format: int64
      required: [taskName]
    TaskUpdateDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        taskName:
          type: string
        taskType:
          type: string
        priority:
          type: integer
        planTime:
          type: string
          format: date-time
      required: [taskId]
    TaskAssignDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        executorId:
          type: integer
          format: int64
        remark:
          type: string
      required: [taskId, executorId]
    TaskAcceptDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
      required: [taskId]
    TaskRejectDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        reason:
          type: string
      required: [taskId]
...