# src/main/java/lizhuoer/agri/agri_system/module/task/domain/dto/TaskUpdateDTO.java
package lizhuoer.agri.agri_system.module.task.domain.dto;

import lombok.Data;

import java.time.LocalDateTime;

@Data
public class TaskUpdateDTO {
    private Long taskId;
    private Long batchId;
    private String taskName;
    private String taskType;
    private Integer priority;
    private LocalDateTime planTime;
}